login and signup, plus cookies!!
Now, each time the user logs in their token will be stored in their user
object in the database as well as the cookie of the browser. When making
any api calls, the token must be present in the cookie. There is no
longer a stored front end user object. Finally, when making API calls,
if all you need is the user object no arguments are needed because the
token is a unique key to any user in the database and it can be accessed
straight from the cookie.